requirements.txt
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
email-validator
python-docx
streamlit
openpyxl